requirements.txt
discord.py==2.3.2
configparser==6.0.0